Main changes to model estimation, LOOCV, cross-correlation, simulations and visu
## Loocv:

loocv consolidation to handle backward models, single models and condition based models
new loocv_func that does loocv correctly: it applies a provided func to the data of n-1 subjects, and then calculates the likelihood on the left out subject. This function can be as complex as you want and also test multiple models at once. It's also not much slower in the end.

## Cross correlation:
fixed using numpy crosscorr method. Works now independent of template width and shape
In addition, while doing these a bunch of small fixes which only happen in edge cases, in particular:

## Simulations
- added sampling_frequency argument to simulate a given sf, defaults to 100Hz.
- removed the location argument as this was conflicting with the declaration of a location in the provided scipy distribution

## Plotting
Changed the computation of electrode values at the different onsets to capture the weigthed mean of the expected shape instead of the mid-point of the template

## Changes to model estimation
- Removed the location argument in the PDF, instead of a statistical location we moved to a censoring of samples following an event. This is mainly build in because samples just next to an event attract to much weight and squashes all n_events into the main one. Adding a padding avoids this situation. The new feature is used as the previous location parameter
- Switched back from PDF to CDF in the probability_distribution function
- Error control on EM and PDF by remove nans on pmfs, min_iteration >= 1 instead of >, and checking for -inf in EM
- minor fixes in the fit() function

## MISC
Name changes for convention